rename /Password directory to AuthPassword.
